+++ activesupport/lib/active_support/notifications/instrumenter.rb
@@ -19,7 +19,7 @@ module ActiveSupport
         begin
           yield payload
         rescue Exception => e
-          payload[:exception] = [e.class.name, e.message]
+          payload[:exception] = [e.class.name, e.message, e.backtrace]
           raise e
         ensure
           finish name, payload