+++ activesupport/test/notifications_test.rb
@@ -209,7 +209,12 @@ module Notifications
     def test_instrument_publishes_when_exception_is_raised
       begin
         instrument(:awesome, :payload => "notifications") do
-          raise "FAIL"
+          begin
+            raise "FAIL"
+          rescue RuntimeError => e
+            e.set_backtrace(['path/to/file/with/error.rb'])
+            raise e
+          end
         end
       rescue RuntimeError => e
         assert_equal "FAIL", e.message
@@ -217,7 +222,7 @@ module Notifications
 
       assert_equal 1, @events.size
       assert_equal Hash[:payload => "notifications",
-        :exception => ["RuntimeError", "FAIL"]], @events.last.payload
+        :exception => ["RuntimeError", "FAIL", ['path/to/file/with/error.rb']]], @events.last.payload
     end
 
     def test_event_is_pushed_even_without_block